Make Drag/Drop a Page concern, not a TermControl concern
